+++ package.json
@@ -103,7 +103,7 @@
     "stylelint-config-recommended-vue": "^1.6.0",
     "stylelint-config-standard-scss": "^13.1.0",
     "typescript": "^5.8.2",
-    "vite": "^5.4.16",
+    "vite": "^6.2.5",
     "vite-plugin-style-import": "^2.0.0",
     "vite-plugin-stylelint": "^4.3.0",
     "vite-plugin-vue-devtools": "^7.7.2",
